WS9: honor image tag in published-images overlay; f64 temperature; doc chain
- install.sh/install.ps1: also write KNOWLEDGE_VERSION (= the operator's tag)
  to .env. The base compose keys image refs on KNOWLEDGE_IMAGE_TAG, but the
  published-images overlay (docker-compose.images.yml) keys on KNOWLEDGE_VERSION;
  the installer layers both, so a tag passed only as KNOWLEDGE_IMAGE_TAG was
  silently dropped for the gateway/substrate images (they fell back to :latest).
  Writing both pins the whole stack. Verified: docker compose config now resolves
  gateway/substrate to the requested tag. (Kept the overlay reading
  KNOWLEDGE_VERSION since README.md and docs/operator/deployment-guide.md
  document that knob.)
- managed_cloud.rs: DEFAULT_MANAGED_TEMPERATURE f32 -> f64 so the request body
  serializes as 0.1 rather than 0.10000000149011612 (f32->f64 widening). Verified.
- lib.rs: crate-level doc chain now reads MLX -> llama.cpp -> ManagedCloud ->
  Fallback (was stale three-adapter chain).

+++ crates/inference_router/src/adapters/managed_cloud.rs
@@ -316,7 +316,7 @@ mod http_client {
 
     /// Default sampling temperature. Synthesis is closer to
     /// extraction than creative generation, so we keep it low.
-    pub const DEFAULT_MANAGED_TEMPERATURE: f32 = 0.1;
+    pub const DEFAULT_MANAGED_TEMPERATURE: f64 = 0.1;
 
     /// Default model name used when `KNOWLEDGE_MANAGED_INFERENCE_MODEL`
     /// is unset. A small, cheap, widely-available OpenAI model is a
